+++ treestamps/tree/dump.py
@@ -1,8 +1,8 @@
 """Dump Methods."""
 
-from collections.abc import MutableMapping
 from contextlib import suppress
 from pathlib import Path
+from warnings import warn
 
 from ruamel.yaml import StringIO
 from termcolor import cprint
@@ -16,17 +16,6 @@ class TreestampsDump(TreestampsInit):
     def _get_relative_path_str(self, abs_path: Path) -> str:
         """Get the path string relative to the root_dir."""
         return str(abs_path.relative_to(self.root_dir))
-
-    def _serialize_timestamps(self):
-        """Dumpable timestamp paths need to be strings."""
-        dumpable_timestamps = {}
-        for abs_path, timestamp in self._timestamps.items():
-            try:
-                rel_path_str = self._get_relative_path_str(abs_path)
-                dumpable_timestamps[rel_path_str] = timestamp
-            except Exception as exc:
-                cprint(f"WARNING: Serializing {abs_path}: {exc}", "yellow")
-        return dumpable_timestamps
 
     def _get_dumpable_program_config(self) -> dict:
         """Set the config tag in the yaml to be dumped."""
@@ -49,15 +38,19 @@ class TreestampsDump(TreestampsInit):
             self._wal.close()
         self._wal = None
 
-    def _prepare_dump(self, yaml: MutableMapping):
+    def dump_dict(self) -> dict:
+        """Seriailiez timestamps and dump to a dict."""
+        yaml = {}
+        for abs_path, timestamp in self._timestamps.items():
+            try:
+                rel_path_str = self._get_relative_path_str(abs_path)
+                yaml[rel_path_str] = timestamp
+            except Exception as exc:
+                cprint(f"WARNING: Serializing {abs_path}: {exc}", "yellow")
         config_yaml = self._get_dumpable_program_config()
         yaml.update(config_yaml)
         self._close_wal()
-
-    def _dumpf(self, yaml: MutableMapping, path: Path):
-        """Dump to file."""
-        self._prepare_dump(yaml)
-        self._YAML.dump(yaml, path)
+        return yaml
 
     def cleanup_old_timestamps(self):
         """Cleanup old timestamps from the disk."""
@@ -71,20 +64,26 @@ class TreestampsDump(TreestampsInit):
                 cprint(f"WARNING: removing old timestamp: {exc}", "yellow")
         self._consumed_paths = set()
 
+    def dumps(self) -> str:
+        """Dump to string."""
+        # NOTE Does not cleanup old timestamps from disk
+        yaml = self.dump_dict()
+        with StringIO() as buf:
+            self._YAML.dump(yaml, buf)
+            return buf.getvalue()
+
+    def _dumpf_init_wal(self):
+        """Write a new wal file to disk."""
+        yaml = self.dump_dict()
+        self._YAML.dump(yaml, self._wal_path)
+
     def dumpf(self) -> None:
         """Serialize timestamps and dump to file."""
-        dumpable_timestamps = self._serialize_timestamps()
-        self._dumpf(dumpable_timestamps, self._dump_path)
+        yaml = self.dump_dict()
+        self._YAML.dump(yaml, self._dump_path)
         self.cleanup_old_timestamps()
 
     def dump(self) -> None:
         """Compatibility alias for dumpf()."""
+        warn("Replaced by Treestamps.dumpf()", PendingDeprecationWarning, stacklevel=2)
         self.dumpf()
-
-    def dumps(self):
-        """Dump to string."""
-        yaml = self._serialize_timestamps()
-        self._prepare_dump(yaml)
-        with StringIO() as buf:
-            self._YAML.dump(yaml, buf)
-            return buf.getvalue()